Fix bugs related to templates

+++ backend/tree_api/views.py
@@ -6,6 +6,7 @@ from .serializers import FileContentSerializer
 from . import app_generator
 from . import tree_generator
 from . import json_translator
+from . import templates
 from django.http import HttpResponse
 from django.http import JsonResponse
 import mimetypes
@@ -153,6 +154,7 @@ def create_file(request):
     # Get the file info
     project_name = request.GET.get('project_name', None)
     filename = request.GET.get('filename', None)
+    template = request.GET.get('template', None)
     
     # Make folder path relative to Django app
     folder_path = os.path.join(settings.BASE_DIR, 'filesystem')
@@ -161,9 +163,10 @@ def create_file(request):
     file_path = os.path.join(action_path, filename)
     
     if not os.path.exists(file_path):
-        with open(file_path, 'w') as f:
-            f.write('')  # Empty content
-        return Response({'success': True})
+        if (templates.create_file_from_template(file_path, filename, template)):
+            return Response({'success': True})
+        else:
+            return Response({'success': False, 'message': 'Template does not exist'}, status=400)
     else:
         return Response({'success': False, 'message': 'File already exists'}, status=400)
 
@@ -200,6 +203,10 @@ def save_file(request):
     project_path = os.path.join(folder_path, project_name)
     action_path = os.path.join(project_path, 'actions')
     file_path = os.path.join(action_path, filename)
+
+    # If file doesn't exist simply return
+    if not os.path.exists(file_path):
+        return JsonResponse({'success': False, 'message': 'File does not exist'}, status=404)
 
     try:
         with open(file_path, 'w') as f:
